cpu/lpc2387: clean up the platform
 - move clock setup from boards/ to cpu/
 - reduce code duplication
