+++ boards/avsextrem/board_init.c
@@ -23,58 +23,9 @@
  */
 #include "cpu.h"
 
-#define CL_CPU_DIV    4
-
-
 /*---------------------------------------------------------------------------*/
-static inline void
-pllfeed(void)
+void board_init(void)
 {
-    PLLFEED = 0xAA;
-    PLLFEED = 0x55;
-}
-
-/*---------------------------------------------------------------------------*/
-void init_clks1(void)
-{
-    // Disconnect PLL
-    PLLCON &= ~0x0002;
-    pllfeed();
-
-    while (PLLSTAT & BIT25);  /* wait until PLL is disconnected before
-                               * disabling - deadlock otherwise */
-
-    // Disable PLL
-    PLLCON &= ~0x0001;
-    pllfeed();
-
-    while (PLLSTAT & BIT24); // wait until PLL is disabled
-
-    SCS |= 0x10;    // main OSC between 15MHz and 24MHz (more stable in tests)
-    SCS |= 0x20;     // Enable main OSC
-
-    while (!(SCS & 0x40));   // Wait until main OSC is usable
-
-    /* select main OSC, 16MHz, as the PLL clock source */
-    CLKSRCSEL = 0x0001;
-
-    // Setting Multiplier and Divider values
-    PLLCFG = 0x0008;          // M=9 N=1 Fcco = 288 MHz
-    pllfeed();
-
-    // Enabling the PLL */
-    PLLCON = 0x0001;
-    pllfeed();
-
-    /* Set clock divider to 4 (value+1) */
-    CCLKCFG = CL_CPU_DIV - 1; // Fcpu = 72 MHz
-}
-
-/*---------------------------------------------------------------------------*/
-void bl_init_ports(void)
-{
-    gpio_init_ports();
-
     /* UART0 */
     PINSEL0 |= BIT4 + BIT6;   // RxD0 and TxD0
     PINSEL0 &= ~(BIT5 + BIT7);
@@ -188,5 +139,4 @@ void bl_init_ports(void)
     PINMODE1 |= (BIT7) | (BIT9) | (BIT11) | (BIT13); // no resistors
     FIO2DIR &= ~(BIT11 + BIT12 + BIT13); //2.11 2.12 2.13 as input
     PINMODE4 |= (BIT23) | (BIT25) | (BIT27); // no resistors
-
 }
