+++ boards/msba2/board_init.c
@@ -29,12 +29,10 @@
 #include "periph/init.h"
 #include "stdio_base.h"
 
-void bl_init_ports(void)
+void board_init(void)
 {
-    gpio_init_ports();
-
     /* UART0 */
-    PINSEL0 |= BIT4 + BIT6;                                 /* RxD0 and TxD0 */
+    PINSEL0 |= BIT4 + BIT6;     /* RxD0 and TxD0 */
     PINSEL0 &= ~(BIT5 + BIT7);
 
     /* LEDS */
@@ -43,39 +41,4 @@ void bl_init_ports(void)
 
     LED0_OFF;
     LED0_OFF;
-
-    /* initialize stdio prior to periph_init() to allow use of DEBUG() there */
-    stdio_init();
-
-    /* trigger static peripheral initialization */
-    periph_init();
-}
-
-void init_clks1(void)
-{
-    /* Disconnect PLL */
-    PLLCON &= ~0x0002;
-    pllfeed();
-
-    /* Disable PLL */
-    PLLCON &= ~0x0001;
-    pllfeed();
-
-    SCS |= 0x20;                        /* Enable main OSC */
-
-    while (!(SCS & 0x40));              /* Wait until main OSC is usable */
-
-    /* select main OSC, 16MHz, as the PLL clock source */
-    CLKSRCSEL = 0x0001;
-
-    /* Setting Multiplier and Divider values */
-    PLLCFG = 0x0008;                    /* M=9 N=1 Fcco = 288 MHz */
-    pllfeed();
-
-    /* Enabling the PLL */
-    PLLCON = 0x0001;
-    pllfeed();
-
-    /* Set clock divider to 4 (value+1) */
-    CCLKCFG = CL_CPU_DIV - 1;           /* Fcpu = 72 MHz */
 }
