+++ cpu/lpc2387/include/cpu.h
@@ -33,6 +33,10 @@ extern uintptr_t __stack_start;     /**< end of user stack memory space */
  */
 void lpc2387_pclk_scale(uint32_t source, uint32_t target, uint32_t *pclksel, uint32_t *prescale);
 
+/**
+ * @brief Initialize lpc2387 cpu clocks
+ */
+void cpu_init_clks(void);
 
 /**
  * @brief install lpc2387 irq
